+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/symbol/intrinsics/GetTypeKindIntrinsic.pas
@@ -1,0 +1,18 @@
+unit GetTypeKindIntrinsic;
+
+interface
+
+implementation
+
+procedure Proc(Arg: TTypeKind); overload;
+begin
+  // Do nothing
+end;
+
+procedure Test;
+begin
+  Proc(GetTypeKind(123));
+  Proc(GetTypeKind(TObject));
+end;
+
+end.